+++ src/litserve/loops.py
@@ -19,9 +19,8 @@ import os
 import pickle
 import sys
 import time
-
 from queue import Empty, Queue
-from typing import List, Tuple, Union
+from typing import Dict, List, Optional, Tuple, Union
 
 from fastapi import HTTPException
 from starlette.formparsers import MultiPartParser
@@ -331,3 +330,43 @@ def run_batched_streaming_loop(
             )
             err_pkl = pickle.dumps(e)
             response_queues[response_queue_id].put((uid, (err_pkl, LitAPIStatus.ERROR)))
+
+
+def inference_worker(
+    lit_api: LitAPI,
+    lit_spec: Optional[LitSpec],
+    device: str,
+    worker_id: int,
+    request_queue: Queue,
+    response_queues: List[Queue],
+    max_batch_size: int,
+    batch_timeout: float,
+    stream: bool,
+    workers_setup_status: Dict[str, bool] = None,
+):
+    lit_api.setup(device)
+    lit_api.device = device
+
+    print(f"Setup complete for worker {worker_id}.")
+
+    if workers_setup_status:
+        workers_setup_status[worker_id] = True
+
+    if lit_spec:
+        logging.info(f"LitServe will use {lit_spec.__class__.__name__} spec")
+    if stream:
+        if max_batch_size > 1:
+            run_batched_streaming_loop(lit_api, lit_spec, request_queue, response_queues, max_batch_size, batch_timeout)
+        else:
+            run_streaming_loop(lit_api, lit_spec, request_queue, response_queues)
+        return
+
+    if max_batch_size > 1:
+        run_batched_loop(lit_api, lit_spec, request_queue, response_queues, max_batch_size, batch_timeout)
+    else:
+        run_single_loop(
+            lit_api,
+            lit_spec,
+            request_queue,
+            response_queues,
+        )
