+++ tests/test_lit_server.py
@@ -26,9 +26,7 @@ from unittest.mock import patch, MagicMock
 import pytest
 
 from litserve.connector import _Connector
-from litserve.server import (
-    inference_worker,
-)
+
 from litserve.server import LitServer
 import litserve as ls
 from fastapi.testclient import TestClient
@@ -54,16 +52,6 @@ def test_device_identifiers(lifespan_mock, simple_litapi):
     # [["cuda:1"], ["cuda:2"]]
     assert server.devices[0][0] == "cuda:1"
     assert server.devices[1][0] == "cuda:2"
-
-
-@patch("litserve.server.run_batched_loop")
-@patch("litserve.server.run_single_loop")
-def test_inference_worker(mock_single_loop, mock_batched_loop):
-    inference_worker(*[MagicMock()] * 6, max_batch_size=2, batch_timeout=0, stream=False)
-    mock_batched_loop.assert_called_once()
-
-    inference_worker(*[MagicMock()] * 6, max_batch_size=1, batch_timeout=0, stream=False)
-    mock_single_loop.assert_called_once()
 
 
 @pytest.mark.asyncio()
